# packages/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rivus API",
    "description": "REST API for Rivus.",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/public/accounts/{slug}": {
      "get": {
        "summary": "Look up an account for the customer self-signup page",
        "tags": [
          "public"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicAccount"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/public/accounts/{slug}/customers": {
      "post": {
        "summary": "Add yourself as a customer of an account",
        "tags": [
          "public"
        ],
        "description": "Backs the self-signup view the scheduling agent links unrecognized senders to. Idempotent by email: if the address already belongs to one of the account’s customers the call succeeds without creating a duplicate — and without revealing that the record already existed.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "default": "",
                    "type": "string",
                    "maxLength": 40
                  },
                  "address": {
                    "default": "",
                    "type": "string",
                    "maxLength": 300
                  }
                },
                "required": [
                  "name",
                  "email"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "slug",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCustomerSignupResult"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
...
}